Assignment 7 - Regression Models
This is an assignment for a data science bootcamp using Google Colaboratory/Jupyter notebooks. This is code copied and pasted over from the CoLab document. Each cell is denoted through the '#NEW CELL' comments.

This code predicts the value of of diamond using "Carat" & "Cut Quality", and later in the code, "Carat" & "Clarity". Using a 13 degree polynomial regression model. I was able to predict prices with a mean absolute error of 223 out of a range of prices from $300-$5000.